requirements.txt
# Python Requirements for Mirror Platform
# Install with: pip install -r requirements.txt

# ============================================================================
# Core API Framework
# ============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# ============================================================================
# Database & Storage
# ============================================================================
sqlalchemy>=2.0.23  # Optional: for advanced ORM features
alembic>=1.13.0     # Optional: for database migrations

# ============================================================================
# MirrorCore - LLM Adapters
# ============================================================================
# Remote LLM APIs
openai>=1.3.0
anthropic>=0.7.0
httpx>=0.25.0

# Local LLM (optional, requires ~4GB RAM + optional GPU)
llama-cpp-python>=0.2.20

# ============================================================================
# MirrorX Engine - Pattern & Tension Detection
# ============================================================================
numpy>=1.24.0
scikit-learn>=1.3.0

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# ============================================================================
# Testing
# ============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# ============================================================================
# Code Quality
# ============================================================================
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# ============================================================================
# Optional: Enhanced Features
# ============================================================================
# Redis for caching (production)
# redis>=5.0.0

# Celery for background tasks (production)
# celery>=5.3.0

# Monitoring (production)
# prometheus-client>=0.19.0

# Websockets for real-time updates
websockets>=12.0

# Rate limiting
slowapi>=0.1.9

# ============================================================================
# Installation Notes
# ============================================================================
# 
# Minimal Install (API + Remote LLM only):
#   pip install fastapi uvicorn pydantic openai anthropic
#
# Local LLM (Privacy-First, requires 4GB+ RAM):
#   pip install llama-cpp-python
#   # For GPU support (CUDA):
#   CMAKE_ARGS="-DLLAMA_CUBLAS=on" pip install llama-cpp-python
#
# Development:
#   pip install pytest black isort
#
# Production:
#   pip install redis celery prometheus-client
#
